refactor(install): move hook-command quoting into install/hook_cmd/{unix,windows}
Localize the OS halves (shell_single_quote / windows cmd-safety + 8.3 FFI)
behind the hook/-style sibling-module split, re-exported as shell_hook_command.
Keeps the injected-short_path pure core (tests run on macOS). No behavior change.
